+++ Tests/TemplatesTests/ColorsTests.swift
@@ -27,25 +27,22 @@ class ColorsTests: XCTestCase {
   }
 
   func testDefault() {
-    test(template: "colors-default",
+    test(template: "default",
          contextNames: Contexts.all,
-         outputPrefix: "default",
          directory: .colors,
          contextVariations: variations)
   }
 
   func testRawValue() {
-    test(template: "colors-rawValue",
+    test(template: "rawValue",
          contextNames: Contexts.rawOnly,
-         outputPrefix: "rawValue",
          directory: .colors,
          contextVariations: variations)
   }
 
   func testSwift3() {
-    test(template: "colors-swift3",
+    test(template: "swift3",
          contextNames: Contexts.all,
-         outputPrefix: "swift3",
          directory: .colors,
          contextVariations: variations)
   }