+++ Tests/TemplatesTests/ImagesTests.swift
@@ -26,41 +26,36 @@ class ImagesTests: XCTestCase {
   }
 
   func testDefault() {
-    test(template: "images-default",
+    test(template: "default",
          contextNames: Contexts.all,
-         outputPrefix: "default",
          directory: .images,
          contextVariations: variations)
   }
 
   func testAllValues() {
-    test(template: "images-allvalues",
+    test(template: "allvalues",
          contextNames: Contexts.all,
-         outputPrefix: "allvalues",
          directory: .images,
          contextVariations: variations)
   }
 
   func testSwift3() {
-    test(template: "images-swift3",
+    test(template: "swift3",
          contextNames: Contexts.all,
-         outputPrefix: "swift3",
          directory: .images,
          contextVariations: variations)
   }
 
   func testDotSyntax() {
-    test(template: "images-dot-syntax",
+    test(template: "dot-syntax",
          contextNames: Contexts.all,
-         outputPrefix: "dot-syntax",
          directory: .images,
          contextVariations: variations)
   }
 
   func testDotSyntaxSwift3() {
-    test(template: "images-dot-syntax-swift3",
+    test(template: "dot-syntax-swift3",
          contextNames: Contexts.all,
-         outputPrefix: "dot-syntax-swift3",
          directory: .images,
          contextVariations: variations)
   }