+++ Tests/TemplatesTests/StoryboardsMacOSTests.swift
@@ -46,26 +46,29 @@ class StoryboardsMacOSTests: XCTestCase {
   }
 
   func testDefault() {
-    test(template: "storyboards-osx-default",
+    test(template: "osx-default",
          contextNames: Contexts.all,
          outputPrefix: "default",
-         directory: .storyboardsMacOS,
+         directory: .storyboards,
+         resourceDirectory: .storyboardsMacOS,
          contextVariations: variations)
   }
 
   func testSwift3() {
-    test(template: "storyboards-osx-swift3",
+    test(template: "osx-swift3",
          contextNames: Contexts.all,
          outputPrefix: "swift3",
-         directory: .storyboardsMacOS,
+         directory: .storyboards,
+         resourceDirectory: .storyboardsMacOS,
          contextVariations: variations)
   }
 
   func testLowercase() {
-    test(template: "storyboards-osx-lowercase",
+    test(template: "osx-lowercase",
          contextNames: Contexts.all,
          outputPrefix: "lowercase",
-         directory: .storyboardsMacOS,
+         directory: .storyboards,
+         resourceDirectory: .storyboardsMacOS,
          contextVariations: variations)
   }
 }